+++ src/components/RedirectIfNoSave/index.ts
@@ -1,0 +1,2 @@
+import RedirectIfNoSave from "./RedirectIfNoSave";
+export default RedirectIfNoSave;